# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RentFlow API
  description: |
    Multi-tenant Rent Management SaaS API.

    **Rules enforced at every endpoint:**
    - `public_id` is used in all URLs and responses — internal `id` is never exposed
    - All resources are scoped to the authenticated owner via JWT
    - Soft-deleted resources (`is_active=false`) are treated as not found (404)
    - All monetary values are numbers (`NUMERIC(10,2)` precision)
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # --------------------------------------------------------------
  # Reusable path parameters
  # --------------------------------------------------------------
  parameters:
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    apartment_public_id:
      name: apartment_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    tenant_public_id:
      name: tenant_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Authentication required"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Access denied"
    NotFound:
      description: Resource not found or soft-deleted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Not found"
    BadRequest:
      description: Business logic error (e.g. duplicate unit, conflicting dates, default category delete)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'

  # --------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        message:
          type: string
    ValidationErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        message:
          type: string
          example: "Validation failed"
    Pagination:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          description: Total number of matching records
        page:
          type: integer
        page_size:
          type: integer

    # ── Auth ───────────────────────────────────────────────────
...
    TenantCreate:
      type: object
      required: [full_name, phone, move_in_date, initial_rent_amount, agreement_start_date]
      properties:
        full_name:
          type: string
          maxLength: 255
        phone:
          type: string
          maxLength: 20
        nid_number:
          type: string
          maxLength: 50
          nullable: true
        address:
          type: string
          nullable: true
        member_count:
          type: integer
          minimum: 1
          default: 1
        move_in_date:
          type: string
          format: date
        initial_rent_amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: "Rent for the initial agreement"
        agreement_start_date:
          type: string
          format: date
          description: "Start date for the initial agreement (typically same as move_in_date)"
    TenantUpdate:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 255
        phone:
          type: string
          maxLength: 20
        nid_number:
          type: string
          maxLength: 50
          nullable: true
        address:
          type: string
          nullable: true
        member_count:
          type: integer
          minimum: 1
    TenantMoveOutRequest:
      type: object
      required: [move_out_date]
      properties:
        move_out_date:
          type: string
          format: date
    TenantData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        apartment_public_id:
          type: string
          format: uuid
        full_name:
          type: string
        phone:
          type: string
        nid_number:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        member_count:
          type: integer
        move_in_date:
          type: string
          format: date
        move_out_date:
          type: string
          format: date
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Tenant Agreements ──────────────────────────────────────
...
    PaymentRecordData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        due_public_id:
          type: string
          format: uuid
        amount_paid:
          type: number
        paid_on:
          type: string
          format: date
        note:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    OverdueItem:
      type: object
      properties:
        tenant_public_id:
          type: string
          format: uuid
        tenant_name:
          type: string
        apartment_unit:
          type: string
        building_name:
          type: string
        overdue_months:
          type: array
          items:
            type: object
            properties:
              due_public_id:
                type: string
                format: uuid
              month:
                type: integer
              year:
                type: integer
              remaining_balance:
                type: number
              status:
                type: string
                enum: [unpaid, partial]
        total_outstanding:
          type: number

    # ── Reports ────────────────────────────────────────────────
    PaymentHistoryEntry:
      type: object
      properties:
        due_public_id:
          type: string
          format: uuid
        month:
          type: integer
        year:
          type: integer
        rent_amount:
          type: number
        total_due:
          type: number
        amount_paid:
          type: number
        remaining_balance:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        due_date:
          type: string
          format: date
          nullable: true
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRecordData'

# ================================================================
# PATHS
# ================================================================
paths:
...
  /apartments/{apartment_public_id}/tenants:
    parameters:
      - $ref: '#/components/parameters/apartment_public_id'
    post:
      tags: [Tenants]
      summary: Add a tenant to an apartment
      description: |
        Atomically creates the tenant record and their initial agreement.
        Sets apartment status to 'occupied'. Returns 400 if the apartment
        already has an active tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        "201":
          description: Tenant added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/TenantData'
                  message: { type: string, example: "Tenant added" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /apartments/{apartment_public_id}/tenants/{tenant_public_id}:
    parameters:
      - $ref: '#/components/parameters/apartment_public_id'
      - $ref: '#/components/parameters/tenant_public_id'
    get:
      tags: [Tenants]
      summary: Get tenant details
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/TenantData'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tenants]
      summary: Update tenant profile details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/TenantData'
                  message: { type: string, example: "Tenant updated" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Tenants]
      summary: Mark tenant as moved out (soft delete)
      description: |
        Sets tenant is_active=false, records move_out_date, closes active agreements
        (is_active=false), and sets the apartment status to 'vacant'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantMoveOutRequest'
      responses:
        "200":
          description: Tenant moved out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { nullable: true, example: null }
                  message: { type: string, example: "Tenant moved out" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  # ==============================================================
  # TENANT AGREEMENTS
  # ==============================================================
...
  /reports/payment-history:
    get:
      tags: [Reports]
      summary: Full payment ledger for a tenant (all dues + their payment records)
      parameters:
        - name: tenant_public_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PaymentHistoryEntry'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /reports/overdue-list:
    get:
      tags: [Reports]
      summary: All unpaid and partial dues across all of the owner's tenants
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - name: building_public_id
          in: query
          schema:
            type: string
            format: uuid
          description: "Filter to a specific building"
        - name: from_month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: from_year
          in: query
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
      responses:
        "200":
          description: Overdue dues list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/OverdueItem'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'